requirements.txt
pycryptodome == 3.10.1 
cryptography == 3.4.7
sockets == 1.0.0
win32gui == 221.6
requests == 2.25.1
Flask == 2.0.1

